fix(taxonomy): audit remediation across reference-data/entity-merge/taxonomy/geocoding
- reference-data: brand ReferenceDataCreateRequest/UpdateRequest validFrom/validTo
  as ISODateString | null (was plain string), matching the response DTO
- react-ui-reference-data: route the row-actions aria-label through t()
  via a new Actions.Menu key (en + fr) instead of a hardcoded English string
- entity-merge: move MergeErrorKind / ClassifiedMergeError type declarations
  into src/types/index.ts (canonical core types location); helpers.ts imports
  them; barrel re-exports the same names (non-breaking)
- react-taxonomy / react-geocoding: relocate buildTaxonomyQueryKey /
  buildGeocodingQueryKey into hooks/query-keys.ts (matching react-entity-merge);
  same symbols re-exported, keys behavior-identical
