+++ packages/@granit/entity-merge/src/helpers.ts
@@ -1,4 +1,10 @@
-import type { FieldConflict, MergeFieldChoices, WinnerSide } from './types/index';
+import type {
+  ClassifiedMergeError,
+  FieldConflict,
+  MergeErrorKind,
+  MergeFieldChoices,
+  WinnerSide,
+} from './types/index';
 
 /**
  * Generate a fresh idempotency key for a merge submission. Prefers
@@ -43,19 +49,6 @@ export function seedFieldChoices(
  */
 export function resolveWinner(conflict: FieldConflict, choices: MergeFieldChoices): WinnerSide {
   return choices[conflict.fieldPath] ?? conflict.default;
-}
-
-/** Coarse classification of a failed merge request, independent of i18n. */
-export type MergeErrorKind = 'conflict' | 'domain' | 'notFound' | 'validation' | 'unknown';
-
-/** Result of {@link classifyMergeError}. */
-export interface ClassifiedMergeError {
-  /** Stable kind the UI maps to a localized message. */
-  readonly kind: MergeErrorKind;
-  /** RFC 7807 `detail`/`title` from the ProblemDetails body, if any. */
-  readonly detail: string | null;
-  /** HTTP status code, when available. */
-  readonly status: number | null;
 }
 
 /**
