+++ packages/@granit/react-geocoding/src/providers/geocoding-provider.tsx
@@ -40,16 +40,3 @@ export const useGeocodingConfig = useConfig;
  * when geocoding is not wired up.
  */
 export const useOptionalGeocodingConfig = useOptionalConfig;
-
-/**
- * Builds a consistent React Query key for geocoding operations. Accepts a `null`
- * config (no provider in scope) so callers can build a stable key for a disabled
- * query without duplicating the default prefix.
- */
-export function buildGeocodingQueryKey(
-  config: GeocodingConfig | null,
-  ...segments: readonly unknown[]
-): readonly unknown[] {
-  const prefix = config?.queryKeyPrefix ?? ['geocoding'];
-  return [...prefix, ...segments];
-}
