+++ packages/@granit/react-taxonomy/src/hooks/use-document-tags.ts
@@ -2,7 +2,9 @@ import { attachTagToDocument, detachTagFromDocument, listDocumentTags } from '@g
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
 import { logMutationError } from '../logger';
-import { buildTaxonomyQueryKey, useTaxonomyConfig } from '../providers/taxonomy-provider';
+import { useTaxonomyConfig } from '../providers/taxonomy-provider';
+
+import { buildTaxonomyQueryKey } from './query-keys';
 
 import type { TagResponse } from '@granit/taxonomy';
 import type { QueryClient, UseMutationResult, UseQueryResult } from '@tanstack/react-query';
