+++ packages/@granit/react-taxonomy/src/hooks/use-tag-mutations.ts
@@ -2,7 +2,9 @@ import { assignTag, createTag, deleteTag, unassignTag, updateTag } from '@granit
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { logMutationError } from '../logger';
-import { buildTaxonomyQueryKey, useTaxonomyConfig } from '../providers/taxonomy-provider';
+import { useTaxonomyConfig } from '../providers/taxonomy-provider';
+
+import { buildTaxonomyQueryKey } from './query-keys';
 
 import type { ResolvedTaxonomyConfig } from '../providers/taxonomy-provider';
 import type {
