+++ packages/@granit/react-ui-reference-data/src/components/reference-data-columns.tsx
@@ -111,7 +111,7 @@ export function createReferenceDataColumns({
                 variant="ghost"
                 size="icon"
                 className="h-8 w-8"
-                aria-label={`Actions for ${entry.labelEn}`}
+                aria-label={t(`${i18nPrefix}.Actions.Menu`, { name: entry.labelEn })}
               >
                 <MoreHorizontal className="h-4 w-4" />
               </Button>
